# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "This is the API documentation for a Todo app. It allows you to perform Basic CRUD operations on todos with filtering, sorting, and searching on all the todos",
        "contact": {
            "email": "andohfrancis9187@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "List todos with filtering, sorting, and searching",
                "operationId": "dd687d49d7f86d473ce56a7806776fc4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter todos by status (completed, in_progress, not_started)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "completed",
                                "in_progress",
                                "not_started"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search todos by title or details",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort todos by a field (title, status, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of todos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Todo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "operationId": "6f5da079c2dace66df52625276205a32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "details": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "completed",
                                            "in_progress",
                                            "not_started"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update an existing todo",
                "operationId": "0e5cec090fe772f5dcb769dd9e38717c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "details": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "completed",
                                            "in_progress",
                                            "not_started"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated todo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete an existing todo",
                "operationId": "2f51ac0f120b8a6c7a8092bbd18174f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "required": [
                    "title",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "The ID of the todo",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the todo",
                        "type": "string"
                    },
                    "details": {
                        "description": "Additional details of the todo",
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "The current status of the todo",
                        "type": "string",
                        "enum": [
                            "completed",
                            "in_progress",
                            "not_started"
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        }
    ]
}